requirements.txt
django-crispy-forms
django==4.0.3
django-debug-toolbar
mysqlclient
djangorestframework